Point CAPI to a forked version of x509lint (#20)
x509Lint supports the concept of different types of certificates (leaf, intermediate, root) however it is hardcoded to only go down the leaf checking path. We have submitted a pull request to fix this in X509lint, however in the meantime we can point the tool to the forked repo that has the fix.