Allow for a list of CRLs to attempt (#26)
This change modifies the interface from accepting a string that is a single CRL URL to a list of CRL URLs to attempt. The application will resolve down the list, left to right, until a CRL succeeds to download. Download errors are discarded unless all downloads fail, in which case all errors are returned for review purposes.