Resource Consumption Fix (#29)
This change makes it such that Firefox Beta and Firefox Nightly are never executing at the same time to ensure that the host box doesn't run out of resources.

It all switches the build to a debug in order to speed up deployment. The actual application itself did not need such optimizations.